refactor: simplify error handling and integrate toaster notifications in ErrorBoundary component
